Added: Osa4 Progress

+++ Osa4/controllers/blogs.js
@@ -1,27 +1,61 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
+const config = require("../utils/config");
+const jwt = require("jsonwebtoken");
+const logger = require("../utils/logger");
+const middleware = require("../utils/middleware");
+const User = require('../models/user');
 
 blogRouter.get("/", async (request, response) => {
-    Blog.find({}).then(result => {
-        response.json(result);
-    });
+    const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+    response.json(blogs);
 });
 
 blogRouter.post("/", async (request, response) => {
-    const blog = await new Blog({
-        title: request.body.title,
-        author: request.body.author,
-        url: request.body.url,
-        likes: request.body.likes,
-    });
-    
-    const savedBlog = await blog.save();
+    const body = request.body;
+    const token = request.token;
+    const decodedToken = jwt.verify(token, config.SECRET);
+    const user = await User.findById(decodedToken.id);
 
+    if (!(token && decodedToken.id)) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+  
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id,
+    });
+  
+    const savedBlog = await blog.save();
+  
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
+  
     response.status(201).json(savedBlog.toJSON());
 });
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
+    const body = request.body;
+    const token = request.token;
+    const decodedToken = jwt.verify(token, config.SECRET);
+    const user = await User.findById(decodedToken.id);
+
+    if (!(token && decodedToken.id)) {
+      return response.status(401).json({ error: `Unauthorized user` });
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id);
+
+    if (blogToDelete.user._id.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      return response.status(401).json({ error: `Unauthorized user` });
+    }
+
     response.status(204).end();
 });
