+++ Osa4/tests/blog_api.test.js
@@ -3,7 +3,11 @@ const supertest = require('supertest');
 const app = require('../app');
 const helper = require('./test_helper')
 const Blog = require('../models/blog');
+const User = require("../models/user");
+const config = require("../utils/config");
+const bcrypt = require("bcrypt");
 const api = supertest(app);
+const jwt = require("jsonwebtoken");
 
 beforeEach(async () => {
     await Blog.deleteMany({});
@@ -24,7 +28,20 @@ test("id instead of _id", async () => {
     }
 });
 
-test("A valid blog can be added", async () => {
+  let token = null;
+  beforeAll(async () => {
+    await User.deleteMany({});
+
+    const passwordHash = await bcrypt.hash("sekret", 10);
+    const user = new User({ username: "root", passwordHash });
+    await user.save();
+
+    const userForToken = { username: user.username, id: user.id };
+    token = jwt.sign(userForToken, config.SECRET);
+    return token;
+  });
+
+  test("A valid blog can be added", async () => {
     const newBlog = {
       title: "testi2",
       author: "Mikael",
@@ -32,7 +49,7 @@ test("A valid blog can be added", async () => {
       likes: 10,
     };
     
-    await api.post("/api/blogs").send(newBlog).expect(201).expect("Content-Type", /application\/json/);
+    await api.post("/api/blogs").send(newBlog).set({'Authorization': `Bearer ${ token }`}).expect(201).expect("Content-Type", /application\/json/);
     
     const blogsAtEnd = await helper.blogsInDb();
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1);
@@ -66,24 +83,55 @@ test("Title and/or Url missing", async () => {
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length);
 });
 
+test("User adds a blog", async () => {
+  const newBlog = {
+    title: "User adds a blog",
+    author: "User",
+    url: "google.com",
+    likes: 1,
+  };
+
+  await api.post("/api/blogs").set("Authorization", `Bearer ${token}`).send(newBlog).expect(201).expect("Content-Type", /application\/json/);
+
+  const blogsAtEnd = await helper.blogsInDb();
+  expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1);
+
+  const titles = blogsAtEnd.map((blog) => blog.title);
+  expect(titles).toContain("User adds a blog");
+});
+
 describe("Deletion of a Blog", () => {
+  let token = null;
+
+  beforeEach(async () => {
+    await Blog.deleteMany({});
+    await User.deleteMany({});
+
+    const passwordHash = await bcrypt.hash("sekret", 10);
+    const user = new User({ username: "root", passwordHash });
+    await user.save();
+
+    const userForToken = { username: user.username, id: user.id };
+    token = jwt.sign(userForToken, config.SECRET);
+    return token;
+  });
+
   test("Succeeds with status code 204 if id is valid", async () => {
-    const blogToBeDeleted = {
+    const testBlog = {
       title: 'Delete',
       author: 'Deletor',
       url: 'localhost:3003'
     };
 
-    await api.post('/api/blogs').send(blogToBeDeleted).expect(201).expect("Content-Type", /application\/json/);
+    await api.post('/api/blogs').send(testBlog).set({'Authorization': `Bearer ${ token }`}).expect(201).expect("Content-Type", /application\/json/);
 
     const blogsAtStart = await helper.blogsInDb();
-    const blogToDelete = blogsAtStart[1];
-    console.log('blogToDelete: ', blogToDelete);
-  
-    await api.delete(`/api/blogs/${ blogToDelete.id }`).expect(204);
+    const blogToDelete = blogsAtStart.find(blog => blog.title === testBlog.title);
+    console.log(blogsAtStart)
+    await api.delete(`/api/blogs/${ blogToDelete.id }`).set({'Authorization': `Bearer ${ token }`}).expect(204);
   
     const blogsAtEnd = await helper.blogsInDb();
-    expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length);
+    expect(blogsAtEnd).toHaveLength(blogsAtStart.length - 1);
     const blogs = blogsAtEnd.map(blog => blog.title);
     expect(blogs).not.toContain(blogToDelete.title);
   });
